requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes -o requirements.txt
aenum==3.1.16
    # via deepgram-sdk
aiofiles==24.1.0
    # via
    #   deepgram-sdk
    #   pipecat-ai
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   cartesia
    #   deepgram-sdk
    #   pipecat-ai
    #   pipecatcloud
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
async-timeout==5.0.1 ; python_full_version < '3.11'
    # via aiohttp
attrs==25.4.0
    # via
    #   aiohttp
    #   sigtools
audioop-lts==0.2.1 ; python_full_version >= '3.13'
    # via
    #   cartesia
    #   pipecat-ai
cartesia==2.0.17
    # via pipecat-ai
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
click==8.3.1
    # via
    #   nltk
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   loguru
    #   tqdm
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
dataclasses-json==0.6.7
    # via deepgram-sdk
deepgram-sdk==4.7.0
    # via pipecat-ai
deprecation==2.1.0
    # via deepgram-sdk
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via pipecat-ai
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
exceptiongroup==1.3.0 ; python_full_version < '3.11'
    # via anyio
fastapi==0.121.2
    # via
    #   pipecat-ai
    #   pipecat-ai-small-webrtc-prebuilt
    #   pipecatcloud
fastapi-cli==0.0.16
    # via fastapi
fastapi-cloud-cli==0.3.1
    # via fastapi-cli
flatbuffers==25.9.23
    # via onnxruntime
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
future==1.0.0
    # via pyloudnorm
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   cartesia
    #   deepgram-sdk
    #   fastapi
    #   fastapi-cloud-cli
    #   openai
httpx-sse==0.4.0
    # via cartesia
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   yarl
iterators==0.2.0
    # via cartesia
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
jiter==0.12.0
    # via openai
joblib==1.5.2
    # via nltk
llvmlite==0.44.0
    # via numba
loguru==0.7.3
    # via
    #   pipecat-ai
    #   pipecatcloud
markdown==3.10
    # via pipecat-ai
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
nltk==3.9.2
    # via pipecat-ai
numba==0.61.2
    # via
    #   pipecat-ai
    #   resampy
numpy==2.2.6
    # via
    #   numba
    #   onnxruntime
    #   pipecat-ai
    #   pyloudnorm
    #   resampy
    #   scipy
    #   soxr
onnxruntime==1.23.2
    # via pipecat-ai
openai==2.8.1
    # via pipecat-ai
orjson==3.11.4
    # via fastapi
packaging==25.0
    # via
    #   deprecation
    #   marshmallow
    #   onnxruntime
pillow==11.3.0
    # via pipecat-ai
pipecat-ai==0.0.94
    # via exotel-chatbot-dial-out
pipecat-ai-small-webrtc-prebuilt==1.0.0
    # via pipecat-ai
pipecatcloud==0.2.11
    # via exotel-chatbot-dial-out
prompt-toolkit==3.0.52
    # via questionary
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.5
    # via
    #   onnxruntime
    #   pipecat-ai
pydantic==2.12.4
    # via
    #   cartesia
    #   fastapi
    #   fastapi-cloud-cli
    #   openai
    #   pipecat-ai
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via
    #   cartesia
    #   pydantic
pydantic-extra-types==2.10.6
    # via fastapi
pydantic-settings==2.12.0
    # via fastapi
pydub==0.25.1
    # via cartesia
pygments==2.19.2
    # via rich
pyloudnorm==0.1.1
    # via pipecat-ai
pyreadline3==3.5.4 ; sys_platform == 'win32'
    # via humanfriendly
python-dotenv==1.2.1
    # via
    #   pipecat-ai
    #   pipecatcloud
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.3
    # via
    #   fastapi
    #   uvicorn
questionary==2.1.1
    # via pipecatcloud
regex==2025.11.3
    # via nltk
resampy==0.4.3
    # via pipecat-ai
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
scipy==1.15.3 ; python_full_version < '3.11'
    # via pyloudnorm
scipy==1.16.3 ; python_full_version >= '3.11'
    # via pyloudnorm
sentry-sdk==2.44.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sigtools==4.0.1
    # via synchronicity
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soxr==0.5.0.post1
    # via pipecat-ai
starlette==0.49.3
    # via fastapi
sympy==1.14.0
    # via onnxruntime
synchronicity==0.7.7
    # via pipecatcloud
toml==0.10.2
    # via pipecatcloud
tomli==2.3.0 ; python_full_version < '3.11'
    # via fastapi-cli
tqdm==4.67.1
    # via
    #   nltk
    #   openai
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
    #   pipecatcloud
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   cartesia
    #   deepgram-sdk
    #   exceptiongroup
    #   fastapi
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   starlette
    #   synchronicity
    #   typer
    #   typing-inspect
    #   typing-inspection
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.11.0
    # via fastapi
urllib3==2.5.0
    # via sentry-sdk
uvicorn==0.38.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
    #   pipecat-ai
    #   pipecatcloud
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
wait-for2==0.4.1 ; python_full_version < '3.12'
    # via pipecat-ai
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   cartesia
    #   deepgram-sdk
    #   pipecat-ai
    #   uvicorn
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
yarl==1.22.0
    # via aiohttp
